Run examples from each .rst file independently
Sphinx lazily rebuilds, so we only process .rst file which have
changed.  It also seems to process source files in alphabetical
order by filename, so the previous assumption that we'd see all
examples in "book order" was fundamentally flawed.

So add a ":silent:" option to XapianRunExample, and use this to
run any examples which need to have been run before those in each
.rst file.

Based on suggestion by James Aylett.