Fix displayed C++ compile and execute commands
Nesting is used to generated different examples for different languages(The original intention was that you'd be able to download just the examples for a particular language, and so the nesting wouldn't be needed. However since we've neither done that, nor made anything other than the python-baed version available online, this is the sensible thing to do for now.). Changed the displayed compile and execution command lines for c++ examples as they were referring to wrong location for generating the output, wrong location of support.cc and wrong location in the execution command.