Provide working example commands for C++.
The previous ones assumed something about code and data layout that
is simply not true. We've decided that we don't want to provide
downloads of code and data (as we initially planned), so explicitly
guide people in the overview to clone the repository from github,
and adjust the C++ build instructions to match that layout.