Provide working example commands for Java.
This is slightly harder than for other languages, because of the way
that classpaths and JNI work; and because there isn't a standard
place to put third-party libraries.

So now we recommend that Java folks copy the JNI and jar files into
this repository's root, set their CLASSPATH environment variable,
and then run simple commands that should work on that basis.

Our example code is now in the "code.java" package, because that's
its layout on disk; otherwise we have to mess with the classpath
even more to get things to run properly.