Drop defaulting of language in conf.py
We also have code in the makefile to default to python3, and we expect
people to run sphinx via the makefile.  If they don't for some reason
then it's not onerous to have to pass -tpython3.

Dropping this second default means we can now report the situation where
an unknown language is specified, e.g. make html LANGUAGE=C++ (it should
be LANGUAGE=c++).