custom_weighting.rst: Update to reflect 1.5.0 changes
We now document to return a lower case "human" name from the name()
method, as in 1.5.0 this is used as the name when constructing a weight
from a string of name and parameters.

Xapian::TfIdfWeight is now in the API so rename the example simple
tf-idf implementation to PseudoTfIdfWeight.

Note changes to make TradWeight a simple subclass wrapper around
BM25Weight and to deprecate it.